octeontx2-pf: Fix coverity issues
commit a7062541f763485d2d781c0dd414fa75a4667459 from
git@git.assembla.com:cavium/WindRiver.linux.git

 Fix unintended sign extension issues.

Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/94441
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>

+++ drivers/net/ethernet/marvell/octeontx2/nic/otx2_common.c
@@ -650,7 +650,7 @@ int otx2_txschq_config(struct otx2_nic *pfvf, int lvl)
 	} else if (lvl == NIX_TXSCH_LVL_TL4) {
 		parent =  hw->txschq_list[NIX_TXSCH_LVL_TL3][0];
 		req->reg[0] = NIX_AF_TL4X_PARENT(schq);
-		req->regval[0] = parent << 16;
+		req->regval[0] = (u64)parent << 16;
 		req->num_regs++;
 		req->reg[1] = NIX_AF_TL4X_SCHEDULE(schq);
 		req->regval[1] = dwrr_val;
@@ -662,7 +662,7 @@ int otx2_txschq_config(struct otx2_nic *pfvf, int lvl)
 	} else if (lvl == NIX_TXSCH_LVL_TL3) {
 		parent = hw->txschq_list[NIX_TXSCH_LVL_TL2][0];
 		req->reg[0] = NIX_AF_TL3X_PARENT(schq);
-		req->regval[0] = parent << 16;
+		req->regval[0] = (u64)parent << 16;
 		req->num_regs++;
 		req->reg[1] = NIX_AF_TL3X_SCHEDULE(schq);
 		req->regval[1] = dwrr_val;
@@ -675,11 +675,11 @@ int otx2_txschq_config(struct otx2_nic *pfvf, int lvl)
 	} else if (lvl == NIX_TXSCH_LVL_TL2) {
 		parent =  hw->txschq_list[NIX_TXSCH_LVL_TL1][0];
 		req->reg[0] = NIX_AF_TL2X_PARENT(schq);
-		req->regval[0] = parent << 16;
+		req->regval[0] = (u64)parent << 16;
 
 		req->num_regs++;
 		req->reg[1] = NIX_AF_TL2X_SCHEDULE(schq);
-		req->regval[1] = TXSCH_TL1_DFLT_RR_PRIO << 24 | dwrr_val;
+		req->regval[1] = (u64)TXSCH_TL1_DFLT_RR_PRIO << 24 | dwrr_val;
 
 		if (lvl == hw->txschq_link_cfg_lvl && !is_otx2_sdpvf(pfvf->pdev)) {
 			req->num_regs++;
